# swagger.yaml
openapi: 3.0.0
info:
  title: RabbitMQ Flask Service API
  description: API for managing RabbitMQ queues and workers
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
...
paths:
...
  /worker/workers:
    post:
      tags:
        - Worker API
      summary: Find workers for a specific queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queue_name:
                  type: string
      responses:
        '200':
          description: List of workers for the specified queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  queue_name:
                    type: string
                  worker_count:
                    type: integer
                  workers:
                    type: array
                    items:
                      type: object
                      properties:
                        consumer_tag:
                          type: string
                        channel_details:
                          type: object
                        connection_details:
                          type: object
                        log_file:
                          type: object
                          properties:
                            static_url:
                              type: string
                            full_url:
                              type: string
                            view_url:
                              type: string
                        ip_address:
                          type: string
                        pid:
                          type: integer
  /worker/scale/{queue_name}:
    post:
      tags:
        - Worker API
      summary: Scale the number of workers for a queue
      parameters:
        - name: queue_name
          in: path
          required: true
          description: Name of the queue to scale
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                worker_name:
                  type: string
                count:
                  type: integer
      responses:
        '200':
          description: Workers scaled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  queue_name:
                    type: string
                  worker_name:
                    type: string
                  previous_count:
                    type: integer
                  current_count:
                    type: integer
                  target_count:
                    type: integer
                  workers:
                    type: array
                    items:
                      type: object
                      properties:
                        consumer_tag:
                          type: string
                        channel:
                          type: string
                        connection:
                          type: string
                        pid:
                          type: integer
                        worker_status:
                          type: string
                  workers_added:
                    type: integer
                  workers_removed:
                    type: integer
                  status:
                    type: string
                  success:
                    type: boolean
...